+++ crates/microdns-core/src/config.rs
@@ -326,7 +326,7 @@ fn default_recursor_listen() -> String {
     "0.0.0.0:5353".to_string()
 }
 fn default_rest_listen() -> String {
-    "0.0.0.0:8080".to_string()
+    "0.0.0.0:80".to_string()
 }
 fn default_grpc_listen() -> String {
     "0.0.0.0:50051".to_string()
@@ -368,7 +368,7 @@ fn default_peer_dns_port() -> u16 {
     53
 }
 fn default_peer_http_port() -> u16 {
-    8080
+    80
 }
 fn default_peer_grpc_port() -> u16 {
     50051
@@ -452,7 +452,7 @@ default_probe = "ping"
 
 [api.rest]
 enabled = true
-listen = "0.0.0.0:8080"
+listen = "0.0.0.0:80"
 api_key = "secret"
 
 [api.grpc]
@@ -503,7 +503,7 @@ format = "text"
         assert_eq!(config.instance.peers[0].id, "test-g10");
         assert_eq!(config.instance.peers[0].addr, "192.168.10.199");
         assert_eq!(config.instance.peers[0].dns_port, 53); // default
-        assert_eq!(config.instance.peers[0].http_port, 8080); // default
+        assert_eq!(config.instance.peers[0].http_port, 80); // default
         assert_eq!(config.instance.peers[1].dns_port, 5353); // custom
         assert_eq!(config.instance.peers[1].http_port, 9090); // custom
     }